requirements.txt
fastapi
uvicorn
feedparser
pydantic>=2.0.0
pydantic-settings
openai
imagekitio
apscheduler
python-dotenv
feedparser
requests 